requirements.txt
flask
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
opencv-python>=4.5.0
pillow
nltk
python-dotenv
google-generativeai
kaggle
scipy
pyyaml
streamlit